refactor dependency management
